feat(lh-89724): allow appending custom prefixes to microservice APIs
Add a new optional field to services defined in `cloud-fw-mgr-api-docs.config.yml` called
`prefixToAdd`, which specifies a prefix to add to each of the paths in that application. For
example, if your endpoints are accessible on /api/platform/myapp/foos, but the public-api-gateway
makes them available on /api/rest/v1/bars/foos, you should specify /v1/bars as the prefix.